Tidy RHIME runner setup boundaries
