+++ openghg_inversions/rhime/__init__.py
@@ -6,7 +6,7 @@ from openghg_inversions.models.rhime import RhimeModelSpec, SectorSpec
 
 from .params import params_from_config, resolve_flux_sources
 from .runner import RhimeResult, run_rhime, run_rhime_multisector
-from .sampling import RhimeSampler, RhimeSamplingSpec
+from .sampling import RhimeSampler
 from .specs import RhimeOutputSpec, RhimeRunSpec
 
 __all__ = [
@@ -14,7 +14,6 @@ __all__ = [
     "RhimeModelSpec",
     "RhimeOutputSpec",
     "RhimeSampler",
-    "RhimeSamplingSpec",
     "RhimeRunSpec",
     "RhimeResult",
     "params_from_config",
