+++ openghg_inversions/rhime/outputs.py
@@ -1,0 +1,327 @@
+"""RHIME output construction and persistence helpers."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Any, cast
+
+import arviz as az
+import numpy as np
+import xarray as xr
+
+from openghg_inversions.array_ops import sparse_xr_dot
+from openghg_inversions.inversion_data import RhimePreparedInputs
+from openghg_inversions.models import RhimeModelSpec
+from openghg_inversions.postprocessing.inversion_output import InversionOutput
+from openghg_inversions.rhime.specs import RhimeOutputSpec, RhimeRunSpec
+from openghg_inversions.utils import ncdf_encoding
+
+
+@dataclass(frozen=True)
+class RhimeOutputBundle:
+    """Outputs and metadata generated by RHIME output helpers."""
+
+    inv_out: InversionOutput | None = None
+    outputs: dict[str, Any] = field(default_factory=dict)
+    output_metadata: dict[str, Any] = field(default_factory=dict)
+
+
+def _resolve_output_path(
+    save_setting: str | Path | bool, output_path: str | None, filename: str
+) -> Path | None:
+    """Resolve an optional output path from a bool/path save setting."""
+    if not save_setting:
+        return None
+    if isinstance(save_setting, str | Path):
+        return Path(save_setting)
+    if output_path is None:
+        raise ValueError("An output path is required when saving RHIME artifacts.")
+    return Path(output_path) / filename
+
+
+def _define_output_filename(
+    output_path: str | Path,
+    species: str,
+    domain: str,
+    output_name: str,
+    start_date: str,
+    *,
+    ext: str = ".nc",
+) -> Path:
+    """Create the RHIME output filename used for derived NetCDF products."""
+    return Path(output_path) / f"{output_name}_{species}_{domain}_{start_date}{ext}"
+
+
+def _save_inferencedata(idata: az.InferenceData, path: str | Path) -> None:
+    """Save InferenceData, preferring the h5netcdf backend with fallbacks."""
+    failures = []
+    for engine in ("h5netcdf", None, "netcdf4"):
+        try:
+            if engine is None:
+                idata.to_netcdf(str(path), compress=True)
+            else:
+                idata.to_netcdf(str(path), engine=engine, compress=True)
+        except Exception as exc:
+            engine_name = "arviz-default" if engine is None else engine
+            failures.append(f"{engine_name}: {exc}")
+        else:
+            return
+
+    joined_failures = "\n".join(failures)
+    raise RuntimeError(
+        f"Could not save RHIME trace to {path}. Tried h5netcdf, ArviZ default, and netcdf4:\n{joined_failures}"
+    )
+
+
+def _materialise_basis_and_flux_for_output(
+    prepared: RhimePreparedInputs,
+) -> tuple[xr.DataArray, xr.DataArray]:
+    """Return materialised arrays for transitional output adapters.
+
+    TODO(#383/#429): postprocessing should consume ``BasisFunctions`` directly
+    rather than requiring the runner to materialise a flat basis and flux.
+    """
+    basis = prepared.basis_functions.operator.basis_matrix
+    current_state_dim = prepared.basis_functions.operator.meta.state_dim
+    if current_state_dim != "region":
+        basis = basis.rename({current_state_dim: "region"})
+    region_coord = prepared.inv_inputs.region
+    if region_coord.name is None:
+        raise ValueError("prepared.inv_inputs.region must be named so it can be used for reindexing.")
+    basis = basis.reindex({region_coord.name: region_coord})
+    return basis, prepared.basis_functions.flux
+
+
+def _make_inversion_output(
+    *,
+    prepared: RhimePreparedInputs,
+    idata: az.InferenceData,
+    start_date: str,
+    end_date: str,
+    species: str,
+    domain: str,
+) -> InversionOutput:
+    """Create an InversionOutput directly from RHIME inputs and InferenceData.
+
+    This is a transitional direct constructor for the modern RHIME path. It is
+    deliberately not routed through the fixed-basis/inferpymc legacy adapter.
+    This should be refactored when issue #401 defines the modern
+    ``InversionOutput`` contract.
+    """
+    inv_inputs = prepared.inv_inputs
+    basis, flux = _materialise_basis_and_flux_for_output(prepared)
+    nmeasure = np.arange(inv_inputs.sizes["nmeasure"])
+    site_names = (
+        inv_inputs["site_names"]
+        if "site_names" in inv_inputs
+        else xr.DataArray(list(prepared.sites), dims="nsite")
+    )
+
+    obs_prior_factor = inv_inputs["mf_prior_factor"] if "mf_prior_factor" in inv_inputs else None
+    obs_prior_upper_level_factor = (
+        inv_inputs["mf_prior_upper_level_factor"] if "mf_prior_upper_level_factor" in inv_inputs else None
+    )
+
+    def nmeasure_array(name: str, source: xr.DataArray) -> xr.DataArray:
+        """Create a clean nmeasure DataArray without inherited MultiIndex coords."""
+        result = xr.DataArray(
+            source.values,
+            dims=["nmeasure"],
+            coords={"nmeasure": nmeasure},
+            name=name,
+        )
+        result.attrs = source.attrs
+        return result
+
+    return InversionOutput(
+        obs=nmeasure_array("Yobs", inv_inputs["mf"]),
+        obs_err=nmeasure_array("Yerror", inv_inputs["mf_error"]),
+        obs_repeatability=nmeasure_array("Yerror_repeatability", inv_inputs["mf_repeatability"]),
+        obs_variability=nmeasure_array("Yerror_variability", inv_inputs["mf_variability"]),
+        obs_prior_factor=(
+            nmeasure_array("Yobs_prior_factor", obs_prior_factor) if obs_prior_factor is not None else None
+        ),
+        obs_prior_upper_level_factor=(
+            nmeasure_array("Yobs_prior_upper_level_factor", obs_prior_upper_level_factor)
+            if obs_prior_upper_level_factor is not None
+            else None
+        ),
+        site_indicators=nmeasure_array("site_indicator", inv_inputs["site_indicator"]),
+        flux=flux,
+        basis=basis,
+        trace=idata,
+        site_names=site_names,
+        times=nmeasure_array("times", inv_inputs["time"]),
+        start_date=start_date,
+        end_date=end_date,
+        species=species,
+        domain=domain,
+    )
+
+
+def make_standard_output_bundle(
+    *,
+    output_spec: RhimeOutputSpec,
+    run_spec: RhimeRunSpec,
+    model_spec: RhimeModelSpec,
+    idata: az.InferenceData,
+    prepared: RhimePreparedInputs,
+    country_file: str | None,
+) -> RhimeOutputBundle:
+    """Create and optionally save standard RHIME outputs."""
+    if output_spec.output_format == "none":
+        return RhimeOutputBundle()
+
+    outputs: dict[str, Any] = {}
+    output_metadata: dict[str, Any] = {}
+    inv_out = _make_inversion_output(
+        prepared=prepared,
+        idata=idata,
+        start_date=run_spec.start_date,
+        end_date=run_spec.end_date,
+        species=model_spec.species,
+        domain=model_spec.domain,
+    )
+    outputs["inversion_output"] = inv_out
+
+    trace_path = _resolve_output_path(
+        output_spec.save_trace,
+        output_spec.output_path,
+        f"{output_spec.output_name}{run_spec.start_date}_trace.nc",
+    )
+    if trace_path is not None:
+        trace_path.parent.mkdir(parents=True, exist_ok=True)
+        _save_inferencedata(idata, trace_path)
+        output_metadata["trace_path"] = str(trace_path)
+
+    inv_out_path = _resolve_output_path(
+        output_spec.save_inversion_output,
+        output_spec.output_path,
+        f"{output_spec.output_name}{run_spec.start_date}_inversion_output.nc",
+    )
+    if inv_out_path is not None:
+        inv_out_path.parent.mkdir(parents=True, exist_ok=True)
+        inv_out.save(inv_out_path)
+        output_metadata["inversion_output_path"] = str(inv_out_path)
+
+    if output_spec.output_format == "basic":
+        from openghg_inversions.postprocessing.make_outputs import basic_output
+
+        outputs["basic"] = basic_output(inv_out, country_file=country_file)
+    elif output_spec.output_format == "paris":
+        from openghg_inversions.postprocessing.make_paris_outputs import make_paris_outputs
+
+        obs_avg_period = prepared.averaging_period[0] or "0h"
+        kwargs = output_spec.paris_postprocessing_kwargs or {}
+        flux_outs, conc_outs = make_paris_outputs(
+            inv_out,
+            country_file=country_file,
+            domain=model_spec.domain,
+            obs_avg_period=obs_avg_period,
+            **kwargs,
+        )
+        outputs["paris_flux"] = flux_outs
+        outputs["paris_concentration"] = conc_outs
+
+        if output_spec.output_path is not None:
+            Path(output_spec.output_path).mkdir(parents=True, exist_ok=True)
+            conc_file = _define_output_filename(
+                output_spec.output_path,
+                model_spec.species,
+                model_spec.domain,
+                output_spec.output_name + "_conc",
+                run_spec.start_date,
+                ext=".nc",
+            )
+            flux_file = _define_output_filename(
+                output_spec.output_path,
+                model_spec.species,
+                model_spec.domain,
+                output_spec.output_name + "_flux",
+                run_spec.start_date,
+                ext=".nc",
+            )
+            conc_outs.to_netcdf(
+                conc_file, unlimited_dims=["time"], mode="w", encoding=ncdf_encoding(conc_outs)
+            )
+            flux_outs.to_netcdf(
+                flux_file, unlimited_dims=["time"], mode="w", encoding=ncdf_encoding(flux_outs)
+            )
+            output_metadata["paris_concentration_path"] = str(conc_file)
+            output_metadata["paris_flux_path"] = str(flux_file)
+
+    return RhimeOutputBundle(inv_out=inv_out, outputs=outputs, output_metadata=output_metadata)
+
+
+def _make_multisector_flux_diagnostics(
+    *,
+    idata: az.InferenceData,
+    prepared: RhimePreparedInputs,
+    model_spec: RhimeModelSpec,
+) -> xr.Dataset:
+    """Create provisional sector-aware posterior flux diagnostics.
+
+    TODO(#398/#429): replace this runner-local special case with
+    sector-aware postprocessing once the modern output layer supports multiple
+    sectors and ``BasisFunctions`` reconstruction directly.
+
+    Args:
+        idata: InferenceData returned by RHIME sampling.
+        prepared: Prepared RHIME input object containing retained basis functions.
+        model_spec: Model spec containing sector names and variable suffixes.
+
+    Returns:
+        Dataset containing posterior mean scaling factors, sector posterior flux
+        means, and total posterior flux mean.
+    """
+    basis, flux = _materialise_basis_and_flux_for_output(prepared)
+    posterior_flux = []
+    posterior_scaling = []
+
+    for sector in model_spec.sectors:
+        x_name = f"x_{sector.variable_suffix}"
+        x_mean = cast(Any, idata).posterior[x_name].mean(("chain", "draw"))
+        scale_grid = sparse_xr_dot(basis, x_mean)
+        sector_flux = flux.sel(source=sector.flux_source) if "source" in flux.dims else flux
+        posterior_scaling.append(scale_grid.expand_dims(sector=[sector.name]))
+        posterior_flux.append((scale_grid * sector_flux).expand_dims(sector=[sector.name]))
+
+    scaling = xr.concat(posterior_scaling, dim="sector").rename("posterior_scaling_mean")
+    flux_by_sector = xr.concat(posterior_flux, dim="sector").rename("posterior_flux_mean")
+    total_flux = flux_by_sector.sum("sector").rename("posterior_flux_total_mean")
+    return xr.merge([scaling, flux_by_sector, total_flux])
+
+
+def make_multisector_output_bundle(
+    *,
+    output_spec: RhimeOutputSpec,
+    run_spec: RhimeRunSpec,
+    model_spec: RhimeModelSpec,
+    idata: az.InferenceData,
+    prepared: RhimePreparedInputs,
+) -> RhimeOutputBundle:
+    """Create and optionally save transitional multi-sector RHIME outputs."""
+    diagnostics = _make_multisector_flux_diagnostics(
+        idata=idata,
+        prepared=prepared,
+        model_spec=model_spec,
+    )
+    outputs: dict[str, Any] = {"sector_flux_diagnostics": diagnostics}
+    output_metadata: dict[str, Any] = {}
+
+    if output_spec.output_format == "paris":
+        output_metadata["paris_note"] = (
+            "Multi-sector PARIS schema support is not implemented in issue #398; "
+            "sector-aware modern diagnostics were generated instead."
+        )
+    if output_spec.output_path is not None and output_spec.output_format != "none":
+        Path(output_spec.output_path).mkdir(parents=True, exist_ok=True)
+        diagnostics_path = (
+            Path(output_spec.output_path)
+            / f"{output_spec.output_name}{run_spec.start_date}_sector_flux_diagnostics.nc"
+        )
+        diagnostics.to_netcdf(diagnostics_path, mode="w", encoding=ncdf_encoding(diagnostics))
+        output_metadata["sector_flux_diagnostics_path"] = str(diagnostics_path)
+
+    return RhimeOutputBundle(outputs=outputs, output_metadata=output_metadata)
